agent-core: extract runtime subpath; agent-cli stops being a library
Until now, \`editSnapshot\` and \`summarizeSnapshot\` lived in
\`@anipres/agent-cli\` and \`@anipres/agent-mcp\` reached across to
import them from there — a dependency direction that didn't match
the actual layering. The agent-mcp comment "Used by both the CLI's
summarize subcommand and the MCP tool of the same name" sitting in
agent-cli surfaced the asymmetry: a function consumed equally by
both surfaces should live somewhere neither owns.

Move the pure-library functions into a new
\`@anipres/agent-core/runtime\` subpath:

- runtime/edit-snapshot.ts: editSnapshot + types
- runtime/summarize-snapshot.ts: summarizeSnapshot, formatSnapshotSummary,
  SnapshotSummary
- runtime/install-dom-globals.ts: installDomGlobals (idempotent
  happy-dom shim)
- runtime/index.ts: barrel

agent-core now declares anipres + happy-dom as runtime deps (they
were devDeps for tests; runtime needs them too) and exposes the
new subpath via the \`exports\` field. The existing subpath
isolation (web app uses /client + /react, worker uses /server)
keeps happy-dom out of those bundles.

agent-cli becomes bin-only:
- src/index.ts deleted (no library entry surface)
- src/install-dom-globals.ts deleted (moved to agent-core/runtime)
- edit-command.ts / summarize-command.ts shrink to just the file-I/O
  wrappers (\`runEditCommand\`, \`runSummarizeCommand\`) and
  delegate the actual work to agent-core/runtime
- package.json drops main/types/exports + happy-dom + tldraw +
  anipres; deps reduce to @anipres/agent-core + citty
- bin.ts adds \`process.exit(0)\` after \`runMain(main)\` resolves —
  \`import "anipres"\` registers process-level handles (timers via
  tldraw transitives) that no DOM-side teardown can reach, so a
  single-shot CLI needs an explicit exit. The dispose machinery I
  was building turned out not to actually drain those handles
  (verified via \`process.getActiveResourcesInfo()\`), so it's gone.

agent-mcp:
- imports editSnapshot / summarizeSnapshot from agent-core/runtime
- drops @anipres/agent-cli dep entirely

Plus a small unrelated cleanup: trim verbose comments from
AgentPanel.tsx that just restated what the props/JSX already say.
